feat: add proxy support to the auth flow [gh-0]
